+++ src/components/index.ts
@@ -5,6 +5,7 @@ export { ChatView } from "./ChatView";
 export { DiffView } from "./DiffView";
 export { DiffViewer } from "./DiffViewer";
 export { PermissionModal } from "./PermissionModal";
+export { PermissionPopup } from "./PermissionPopup";
 export { SettingsView } from "./SettingsView";
 export { StatusBar } from "./StatusBar";
 export { AnimatedStatus } from "./AnimatedStatus";